requirements.txt
backcall==0.2.0
click==8.0.1
cloudpickle==1.6.0
cycler==0.10.0
decorator==5.0.9
Flask==2.0.1
Flask-Cors==3.0.10
imbalanced-learn==0.8.0
imblearn==0.0
ipython==7.26.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.2
llvmlite==0.37.0
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
numba==0.54.0
numpy==1.20.3
pandas==1.3.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.20
ptyprocess==0.7.0
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
scikit-learn==0.24.2
scipy==1.7.1
shap==0.39.0
six==1.16.0
slicer==0.0.7
threadpoolctl==2.2.0
tqdm==4.62.2
traitlets==5.0.5
typing-extensions==3.10.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
xlrd==2.0.1